Using llvm-style directory structure & removing confusing file names
