+++ tools/hls-verifier/CMakeLists.txt
@@ -2,16 +2,16 @@ set(LLVM_LINK_COMPONENTS
   Support
 )
 
+include_directories(${DYNAMATIC_SOURCE_DIR}/tools/hls-verifier/include)
+
 add_llvm_tool(hls-verifier
   hls-verifier.cpp
-  src/CAnalyser.cpp
-  src/Help.cpp
-  src/HlsCoVerification.cpp
-  src/HlsLogging.cpp
-  src/HlsVhdlTb.cpp
-  src/HlsVhdlVerification.cpp
-  src/Utilities.cpp
-  src/VerificationContext.cpp
+  lib/CAnalyser.cpp
+  lib/Help.cpp
+  lib/HlsLogging.cpp
+  lib/HlsVhdlTb.cpp
+  lib/Utilities.cpp
+  lib/VerificationContext.cpp
 )
 
 llvm_update_compile_flags(hls-verifier)
